Transpose Matrices
Import:
Matrices are being bound in the wrong orientation, so call for the
matrices to be transposed after being bound (cbind).

Save:
If cbind fails during save.lipds.csv, then try to transpose the matrix
and bind again just in case the matrix happens to be in the wrong
orientation.