Setting global variables
- Whenever a bagit file is located for the first time, assign it to the
global environment so you don’t have to set it again for the same
session.
- Bagit py is included in the package, and should have an absolute path
relative to the package install. System call will try that path first
before prompting the user to locate a bagit file. (This should work in
almost all cases so the user doesn’t need to locate a bagit py)
- Initial working directory is set to global env when the user picks
the directory where their files are stored. This works with the
return.to.root function that resets back to the “home” directory
whenever a function might fail unexpectedly and leave the user in an
unknown tmp directory.