Bugs: Indexing and ensemble numbers
- Ensemble matrices are sometimes labeled as numeric types. However,
they were testing TRUE as a matrix and a numeric, so they never made it
to the ensemble table handling condition. Thus, no “numbers” were ever
ranges. I moved the ensemble condition above the general condition, and
now it works.
- Bug when indexing table columns back to numbered-indices. It was
checking for columns based on variable names, but some columns dont
have variable names and therefore weren’t being re-indexed. This broke
the saving process later because it depends on index-by-number. Changed
the process to not depend on variable names in columns.