Distribution Tables fix
- It appears that distribution tables were never tested, or not tested
well. The “Rustic.lpd” test file proved that it would not import a file
with distribution data. The “load_lipds_merge_data.R” file was really
messing up the data structure during import and crashing it down the
line. It is working now with normal LiPD files, and this new test case.
May need more work when we have more test cases to try out.